src/requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.1
billiard==3.5.0.5
celery==4.2.1
Django==2.2.3
django-filter==2.2.0
django-jet==1.0.8
djangorestframework==3.10.3
gunicorn==19.9.0
kombu==4.3.0
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.7.5
pytz==2019.2
redis==2.10.6
soupsieve==1.9.5
sqlparse==0.3.0
vine==1.3.0
